requirements.txt
pytorch
torchvision
numpy
matplotlib
pillow
tqdm